[refactor] entity 패키지 분리 및 infrastrucutre 패키지로 이동 #11
